refactor(api): remove `DatasetHandlers` move to `handlers.go`
