+++ api/params_test.go
@@ -138,3 +138,55 @@ func TestUnmarshalGetParams(t *testing.T) {
 		})
 	}
 }
+
+func TestParseGetParamsAcceptHeader(t *testing.T) {
+	// Construct a request with "Accept: text/csv"
+	r, _ := http.NewRequest("GET", "/get/peer/my_ds", nil)
+	r.Header.Add("Accept", "text/csv")
+	r = mux.SetURLVars(r, map[string]string{"peername": "peer", "name": "my_ds"})
+	setRefStringFromMuxVars(r)
+	args := &lib.GetParams{}
+	err := UnmarshalParams(r, args)
+	if err != nil {
+		t.Fatal(err)
+	}
+	expectArgs := &lib.GetParams{
+		Refstr:   "peer/my_ds",
+		Selector: "body",
+		Format:   "csv",
+		All:      true,
+	}
+
+	if diff := cmp.Diff(expectArgs, args); diff != "" {
+		t.Errorf("output mismatch (-want +got):\n%s", diff)
+	}
+
+	// Construct a request with format=csv and "Accept: text/csv", which is ok
+	r, _ = http.NewRequest("GET", "/get/peer/my_ds?format=csv", nil)
+	r.Header.Add("Accept", "text/csv")
+	r = mux.SetURLVars(r, map[string]string{"peername": "peer", "name": "my_ds"})
+	setRefStringFromMuxVars(r)
+	args = &lib.GetParams{}
+	err = UnmarshalParams(r, args)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if diff := cmp.Diff(expectArgs, args); diff != "" {
+		t.Errorf("output mismatch (-want +got):\n%s", diff)
+	}
+
+	// Construct a request with format=json and "Accept: text/csv", which is an error
+	r, _ = http.NewRequest("GET", "/get/peer/my_ds?format=json", nil)
+	r.Header.Add("Accept", "text/csv")
+	r = mux.SetURLVars(r, map[string]string{"peername": "peer", "name": "my_ds"})
+	setRefStringFromMuxVars(r)
+	args = &lib.GetParams{}
+	err = UnmarshalParams(r, args)
+	if err == nil {
+		t.Error("expected to get an error, but did not get one")
+	}
+	expectErr := `format "json" conflicts with header "Accept: text/csv"`
+	if expectErr != err.Error() {
+		t.Errorf("error mismatch, expect: %q, got %q", expectErr, err)
+	}
+}
